Load out update!
Added the load out to the beginning of the brawl! After players select
their colours, they choose what items they want equipped, and to what
button.

Also, if you press "back" on a controller or "F" or ";" on the keyboard,
the item icons will be displayed over your character to indicate what
button they are mapped too.

This commit also includes some small tweaks to combat.